refactor thread pool
